# types.gen.go
// Package schema provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.0 DO NOT EDIT.
package schema

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes bearerAuthContextKey = "BearerAuth.Scopes"
)

// Defines values for ChatMessageRole.
const (
	Assistant ChatMessageRole = "assistant"
	User      ChatMessageRole = "user"
)

// Valid indicates whether the value is a known member of the ChatMessageRole enum.
func (e ChatMessageRole) Valid() bool {
	switch e {
	case Assistant:
		return true
	case User:
		return true
	default:
		return false
	}
}

// ChatDetail defines model for ChatDetail.
type ChatDetail struct {
	CreatedAt    time.Time          `json:"created_at"`
	Id           openapi_types.UUID `json:"id"`
	Messages     []ChatMessage      `json:"messages"`
	SystemPrompt *string            `json:"system_prompt,omitempty"`
	UpdatedAt    time.Time          `json:"updated_at"`
}

// ChatMessage defines model for ChatMessage.
type ChatMessage struct {
	Content    string             `json:"content"`
	CreatedAt  time.Time          `json:"created_at"`
	ExternalId *string            `json:"external_id,omitempty"`
	Id         openapi_types.UUID `json:"id"`
	Role       ChatMessageRole    `json:"role"`
}

// ChatMessageRole defines model for ChatMessage.Role.
type ChatMessageRole string

// ChatSummary defines model for ChatSummary.
type ChatSummary struct {
	CreatedAt    time.Time          `json:"created_at"`
	Id           openapi_types.UUID `json:"id"`
	SystemPrompt *string            `json:"system_prompt,omitempty"`
	UpdatedAt    time.Time          `json:"updated_at"`
}

// CreateChatBody defines model for CreateChatBody.
type CreateChatBody struct {
	// SystemPrompt Optional system prompt to guide the assistant
	SystemPrompt *string `json:"system_prompt,omitempty"`
}

// PaginatedChats defines model for PaginatedChats.
type PaginatedChats struct {
	Data          []ChatSummary `json:"data"`
	NextPageToken *string       `json:"next_page_token,omitempty"`
}

// SendMessageBody defines model for SendMessageBody.
type SendMessageBody struct {
	// Content Message content
	Content string `json:"content"`
}

// ChatId defines model for chat_id.
type ChatId = openapi_types.UUID

// PageSize defines model for page_size.
type PageSize = int

// BadRequest defines model for BadRequest.
type BadRequest struct {
	Error string `json:"error"`
}

// NotFound defines model for NotFound.
type NotFound struct {
	Error string `json:"error"`
}

// ServerError defines model for ServerError.
type ServerError struct {
	Error string `json:"error"`
}

// bearerAuthContextKey is the context key for BearerAuth security scheme
type bearerAuthContextKey string

// ListChatsParams defines parameters for ListChats.
type ListChatsParams struct {
	PageSize *PageSize `form:"page_size,omitempty" json:"page_size,omitempty"`
	BeforeId *string   `form:"before_id,omitempty" json:"before_id,omitempty"`
}

// ResearchJSONBody defines parameters for Research.
type ResearchJSONBody struct {
	// Prompt Research prompt
	Prompt string `json:"prompt"`
}

// CreateChatJSONRequestBody defines body for CreateChat for application/json ContentType.
type CreateChatJSONRequestBody = CreateChatBody

// SendChatMessageJSONRequestBody defines body for SendChatMessage for application/json ContentType.
type SendChatMessageJSONRequestBody = SendMessageBody

// ResearchJSONRequestBody defines body for Research for application/json ContentType.
type ResearchJSONRequestBody ResearchJSONBody